requirements.txt
Pillow
pillow-heif
pytz
pythoncom